refactor(schema): rename generated module to schema_generated, add overlay test
Addresses the review of #921/#922:

- Rename guessit/schema.py to guessit/schema_generated.py so the public
  schema() function no longer shadows a submodule of the same name (import
  guessit.schema now unambiguously yields the accessor). Update every
  reference: api.py, __init__.py, tests, scripts/gen_schema.py, docstrings
  and CLAUDE.md. The regenerated module is byte-identical (pure rename).
- Add a unit test for overlay_config_enums covering a property absent from
  the frozen base (custom rules_builder): inferred type, enum vs free-form.

refs #921
